Dashboard/requirements.txt
amqp==2.5.2
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
Automat==20.2.0
awscli==1.18.84
billiard==3.6.3.0
boto3==1.14.7
botocore==1.17.7
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
distlib==0.3.0
Django==3.0.5
django-celery-results==1.2.1
docutils==0.15.2
filelock==3.0.12
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
importlib-resources==1.5.0
incremental==17.5.0
jmespath==0.10.0
kombu==4.6.8
lxml==4.5.0
mysqlclient==1.4.6
numpy==1.18.4
oauthlib==3.1.0
pandas==1.0.3
parsel==1.5.2
pbr==5.4.5
Pillow==7.1.2
# pkg-resources==0.0.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyDispatcher==2.0.5
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
PySocks==1.7.1
python-dateutil==2.8.1
python-scrapyd-api==2.1.2
pytz==2020.1
PyYAML==5.3.1
queuelib==1.5.0
redis==3.5.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==3.4.2
s3transfer==0.3.3
Scrapy==2.1.0
scrapyd==1.2.1
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
stevedore==1.32.0
tweepy==3.8.0
Twisted==20.3.0
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.20
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
w3lib==1.21.0
zipp==3.1.0
zope.interface==5.1.0
demjson
# THIS IS A COOMMENT
